further matchmaker process

+++ matchmaker.coffee
@@ -234,31 +234,31 @@ class Queue
     assign them.
   ###
   assignPlayers: ->
-    #@cleanupQueue()
+    @cleanupQueue()
     dbc.query("SELECT count(*) FROM queue", (error, response) =>
       queueCount = response[0]['count(*)']
+      console.log(queueCount)
       while queueCount >= 2
         dbc.query("SELECT queue.id, CONCAT(players.jid, '/', queue.resource) AS jid
                    FROM queue, players WHERE players.id=queue.user_id    
                    ORDER BY queue.queued_at ASC
                    LIMIT 2", (error, response) =>
           # Assign the two players to each other
-          if response[0].jid and response[1].jid
+          if response.length == 2
             @mm.xmppClient.send new xmpp.Element('message', {'type': 'normal', 'to': response[0].jid})
               .c('battleship', {'xmlns': 'http://battleship.me/xmlns/'})
-              .c('queueing', {'action': 'assign', 'id': response[0].id}).up()
-              .c('partner', {'jid': response[1].jid})
+              .c('queueing', {'action': 'assign', 'jid': response[1]['jid'], 'mid': 42})
             @mm.xmppClient.send new xmpp.Element('message', {'type': 'normal', 'to': response[1].jid})
               .c('battleship', {'xmlns': 'http://battleship.me/xmlns/'})
-              .c('queueing', {'action': 'assign', 'id': response[1].id}).up()
-              .c('partner', {'jid': response[0].jid})
-            console.log("Assigned #{response[0].jid} and #{response[1].jid}")
+              .c('queueing', {'action': 'assign', 'jid': response[0]['jid'], 'mid': 42})
+              
+            console.log("Assigned #{response[0]['jid']} and #{response[1]['jid']}")
+            
+            # delete the queueing entry. maybe this should be done after confirmation
+            # todo
+            dbc.query("DELETE FROM queue WHERE id IN (#{response[0].id},#{response[1].id})")
           else
             console.log("Tried to assign two players, but I got no jids :O")
-          
-          # delete the queueing entry. maybe this should be done after confirmation
-          # todo
-          dbc.query("DELETE FROM queue WHERE id IN (#{response[0].id},#{response[1].id})")
         )
         queueCount -= 2;
     )
